link: use the vendor dir instead of the global workspace
(This patch changes the current API.)

The previous behavior of `gx-go link` linked packages to the global workspace in
`$GOPATH/gx/ipfs/` and affected any packages that would depend on it.

Use the Go vendor directory to link packages locally. That implies an explicit
parent-dependency relationship where a dependency package is now linked to the
vendor directory of a particular parent package (instead of linking it to any
package that might depended on that dependency specific hash). Using the vendor
directory allows the import paths to remain the same (Go will look in that
directory before looking in the `$GOPATH` where the global Gx workspace is
located).